+++ CHANGELOG.md
@@ -6,6 +6,11 @@ All notable changes to this project will be documented in this file.
 
 ---
 
+## [1.0.16] - 2025-03-04
+### Added
+- [Serialization Surrogate] more Unity data types
+- [Json] JsonConverters for more Unity data types
+
 ## [1.0.15] - 2025-03-04
 ### Added
 - [Serialization Surrogate] Bounds and Matrix4x4